fix(sanity): ambient types for custom Vitest matchers
